# backend/app/openapi.yaml
openapi: 3.1.0
info:
  title: Fire Crow API
  description: Agentic Security Intelligence Platform Backend
  version: 1.0.0
paths:
...
  /api/v1/system/status:
    get:
      tags:
      - System
      summary: System Status
      description: Return backend readiness and tenant-scoped operational status for
        the control panel.
      operationId: system_status_api_v1_system_status_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
  /api/v1/system/database/stats:
    get:
      tags:
      - System
      summary: Get Database Stats
      description: Retrieve detailed database metrics, table row counts, and storage
        metrics for admin dashboard.
      operationId: get_database_stats_api_v1_system_database_stats_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
...
components:
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer